refactor: Move parser/util.py into util/*
