+++ main.py
@@ -1,7 +1,7 @@
 import cProfile
 import time
 
-from parser.util import readfile
+from util import readfile
 from parser.cst.treegen import TreeGen
 from parser.lexer import Tokenizer, print_tokens
 from parser.cst.tree_print import tprint
